requirements.txt
Flask==3.0.3
google-cloud-bigquery==3.24.0
python-dotenv==1.0.1
google-cloud-geminidataanalytics==0.1.0